pyproject.toml
[project]
name = "fastapi-mongodb-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beanie>=2.0.1",
    "fastapi[standard]>=0.123.0",
    "pydantic-settings>=2.12.0",
    "pymongo>=4.15.4",
]

[dependency-groups]
auth0 = [
    "pyjwt[crypto]>=2.10.1",
]
dev = [
    "ruff>=0.14.7",
]

[tool.ruff]
lint.select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
lint.ignore = [
    "B008",
]